version 3.3.2-yac2 - Update display information
To provide more useful stats when connecting to a vardiff pool,
accepted, rejected and utility have been updated to track the difficulty
instead of the number of shares (this brings it in line with more recent
versions of cgminer).  Added Found Blocks (FB) to the status summary
section of the interface.